com_members: confine picture lookups, scope the mentions API, close open redirects, CSPRNG activation codes, token-check password change, fix PHP 8 fatals

+++ core/components/com_members/admin/controllers/media.php
@@ -92,7 +92,7 @@ class Media extends AdminController
 
 		// must be in allowed extensions
 		$pathInfo = pathinfo($file['name']);
-		$ext = $pathInfo['extension'];
+		$ext = isset($pathInfo['extension']) ? $pathInfo['extension'] : '';
 		if (!in_array($ext, $allowedExtensions))
 		{
 			$these = implode(', ', $allowedExtensions);
@@ -182,13 +182,14 @@ class Media extends AdminController
 			}
 
 			// Get the file thumbnail name
+			$curthumb = '';
 			if ($file == 'profile.png')
 			{
 				$curthumb = 'thumb.png';
 			}
 
 			// Remove the thumbnail
-			if (file_exists($path . DS . $curthumb))
+			if ($curthumb && file_exists($path . DS . $curthumb))
 			{
 				if (!Filesystem::delete($path . DS . $curthumb))
 				{
